Removed Aggregatables and moved aggregation methods into Collection.
